fix: 查询总记录数返回long

+++ meeleet-sys/meeleet-sys-boot/src/main/java/com/meeleet/cloud/sys/rpc/impl/SysUserRpcServiceImpl.java
@@ -91,7 +91,7 @@ public class SysUserRpcServiceImpl implements ISysUserRpcService {
     }
 
     @Override
-    public int count() {
+    public long count() {
         return sysUserService.count();
     }
